backend/requirements.txt
fastapi
uvicorn
litellm
python-dotenv
requests